Tools: Reverted CLI usage of changelog-generate, refactored upload
