+++ .circleci/config.yml
@@ -342,7 +342,7 @@ jobs:
           command: sudo apt-get install cpio
       - run:
           name: Generate changelog since last tag
-          command: (git describe --abbrev=0 --tags | xargs -I{} node changelog/generate --pr <<# parameters.add_build_metadata >>--build-metadata<</ parameters.add_build_metadata >> --since {}) && node changelog/generate
+          command: (git describe --abbrev=0 --tags | xargs -I{} node changelog/generate --pr <<# parameters.add_build_metadata >>--build-metadata<</ parameters.add_build_metadata >> --since {}) && node changelog/generate-html
       - run:
           name: Changelog - Create git-describe named folder and copy changelog
           command: git ls-files -m --others --exclude-standard | grep -i changelog | cpio -pdm "$(git describe)/"
